fixed fatal error with bound variables

+++ files/lib/system/cronjob/DeleteUnusedToDoCategoriesCronjob.class.php
@@ -50,11 +50,13 @@ class DeleteUnusedToDoCategoriesCronjob extends AbstractCronjob {
 			}
 		}
 		
-		$conditions = new PreparedStatementConditionBuilder();
-		$conditions->add("id IN (?)", array($delete));
-		$sql = "DELETE FROM wcf" . WCF_N . "_todo_category
+		if(!empty($delete)) {
+			$conditions = new PreparedStatementConditionBuilder();
+			$conditions->add("id IN (?)", array($delete));
+			$sql = "DELETE FROM wcf" . WCF_N . "_todo_category
 				".$conditions;
-		$statement = WCF::getDB()->prepareStatement($sql);
-		$statement->execute($conditions->getParameters());
+			$statement = WCF::getDB()->prepareStatement($sql);
+			$statement->execute($conditions->getParameters());
+		}
 	}
 }
